Store files in a file set as we walk the tree
The idea is that we store files in a file set until it is full. The
set is full if the total size of all of its files reaches the maximum
file size, or if the number of files reaches the maximum file count.

When we determine that the set is full, we hand it off to an instance
of the processor class to do something with it. Right now, we just
print information, but eventually there will be a subclass that
calls the git command to actually add the files.