Refactor how files are added as an optimization
While the simplest approach of invoking 'git add' once for each file
will work, this is slow because we have to incur the overhead of
starting a process once for each file. The 'git add' command allows us
to pass multiple files, so we can do so as long as we're careful not
to exceed the maximum number of characters for a command (the
process). I've rewritten the logic for adding files to build batches
of files and then call git add for the whole batch. This results in a
pretty dramatic speedup since we're creating a fraction of the
processes that we used to.

The exact character limit for a command (process) depends on the
system and would probably require testing to determine. I've found a
few sources online that cite a limit of 8,192 characters for Windows
10, so I'm using this value for now. It will allow several files per
batch and doesn't require any system-specific testing.